ФЗ 22. Сервисы: gateway, order, payment, delivery

+++ commerce/interaction-api/src/main/java/ru/practicum/commerce/shoppingstore/dto/BookedProductsDto.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,10 +15,10 @@ import lombok.NoArgsConstructor;
 public class BookedProductsDto {
 
     @NotNull(message = "Необходимо указать общий вес доставки")
-    private Double deliveryWeight;
+    private BigDecimal deliveryWeight;
 
     @NotNull(message = "Необходимо указать общие объём доставки")
-    private Double deliveryVolume;
+    private BigDecimal deliveryVolume;
 
     @NotNull(message = "Необходимо указать есть ли хрупкие вещи в доставке")
     private Boolean fragile;
