+++ commerce/warehouse/src/main/java/ru/practicum/commerce/warehouse/entity/ProductStorageEntity.java
@@ -4,8 +4,13 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -25,13 +30,13 @@ public class ProductStorageEntity {
     private Boolean fragile;
 
     @Column(name = "width", nullable = false)
-    private Double width;
+    private BigDecimal width;
 
     @Column(name = "height", nullable = false)
-    private Double height;
+    private BigDecimal height;
 
     @Column(name = "depth", nullable = false)
-    private Double depth;
+    private BigDecimal depth;
 
     @Column(name = "weight", nullable = false)
     private Double weight;
